+++ superwall/src/main/java/com/superwall/sdk/paywall/request/PaywallRequestManager.kt
@@ -188,7 +188,6 @@ class PaywallRequestManager(
         try {
             val result = storeKitManager.getProducts(
                 paywall.productIds,
-                paywall.name,
                 paywall.products,
                 request.overrides.products
             )
@@ -198,8 +197,7 @@ class PaywallRequestManager(
             val outcome = PaywallLogic.getVariablesAndFreeTrial(
                 result.products,
                 result.productsById,
-                request.overrides.isFreeTrial,
-                { id -> storeKitManager.isFreeTrialAvailable(id) }
+                request.overrides.isFreeTrial
             )
             paywall.productVariables = outcome.productVariables
 //            paywall.swProductVariablesTemplate = outcome.swProductVariablesTemplate