+++ superwall/src/main/java/com/superwall/sdk/store/abstractions/product/StoreProductType.kt
@@ -3,16 +3,10 @@ package com.superwall.sdk.store.abstractions.product
 import java.math.BigDecimal
 import java.util.*
 
-// Define a typealias for List as there is no exact equivalent for Array in Kotlin
-//typealias StoreProductDiscountArray = List<StoreProductDiscount>
-
 interface StoreProductType {
+    val fullIdentifier: String
     val productIdentifier: String
     val price: BigDecimal
-    val subscriptionGroupIdentifier: String?
-
-    //    val swProductTemplateVariablesJson: JsonObject
-//    val swProduct: SWProduct
     val localizedPrice: String
     val localizedSubscriptionPeriod: String
     val period: String
@@ -31,6 +25,7 @@ interface StoreProductType {
     val yearlyPrice: String
     val hasFreeTrial: Boolean
     val localizedTrialPeriodPrice: String
+    val trialPeriodPrice: BigDecimal
     val trialPeriodEndDate: Date?
     val trialPeriodEndDateString: String
     val trialPeriodDays: Int
@@ -47,11 +42,7 @@ interface StoreProductType {
     val currencyCode: String?
     val currencySymbol: String?
     val regionCode: String?
-    // Not relevant for Android
-//    val isFamilyShareable: Boolean
-    /// The period details for products that are subscriptions.
-    /// - Returns: `nil` if the product is not a subscription.
     val subscriptionPeriod: SubscriptionPeriod?
-//    val introductoryDiscount: StoreProductDiscount?
-//    val discounts: StoreProductDiscountArray
+
+    fun trialPeriodPricePerUnit(unit: SubscriptionPeriod.Unit): String
 }